+++ Changelog.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Don't consider compile to i686 on x86_64 Windows cross compiling in [#923](https://github.com/PyO3/maturin/pull/923)
+
 ## [0.12.16] - 2022-05-16
 
 * Add Linux armv7l python sysconfig in [#901](https://github.com/PyO3/maturin/pull/901)